Relocate User and OrganizationMembership table models into the model layer
Standardize on LiteLLM_UserTable, now defined in litellm/models/user.py and
re-exported from litellm.proxy._types. Merge in the schema columns the prior
canonical class omitted (team_id, organization_id, object_permission_id,
password, max_parallel_requests, allowed_cache_controls, policies) so reads
through UserRepository no longer drop data. Relocate the referenced
LiteLLM_OrganizationMembershipTable alongside it. Repoint UserRepository; drop
the hand-rolled User domain model.
